+++ CHANGELOG.md
@@ -9,3 +9,8 @@
 ### Changed
 
 - Move datasets to `sm_datasets/datasets` directory
+
+### Fixed
+
+- Invalid dataset directory
+- Use the `kgns.id_to_uri` function of the provided `kgns` instead of fixed wikidata namespace.